requirements.txt
llama-cpp-python==0.2.11 --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
huggingface-hub==0.19.4
requests==2.31.0
colorama==0.4.6
python-dotenv==1.0.0
rich==13.7.0
flask==3.0.0
flask-cors==4.0.0